requirements.txt
Flask==2.3.3
pymongo[srv]==4.5.0
google-generativeai==0.3.1
groq>=0.11.0
PyPDF2==3.0.1
python-dotenv==1.0.0
dnspython==2.4.2
sentence-transformers==2.6.1
numpy==2.1.2
scikit-learn==1.7.2
huggingface_hub==0.24.6
underthesea
PyMuPDF>=1.23.0  # fitz - để convert PDF thành ảnh
Pillow>=10.0.0  # PIL - để xử lý ảnh
requests>=2.31.0  # để gọi API Vintern
torch>=2.0.0  # để xử lý tensors